增加两种创建生成物素材；布局Level3；增加修改部分文件注释头；统一代码编码为GB2312；消除部分warnings

+++ GeometryRun/GeometryRun/KeyAndObjUpdate.h
@@ -18,7 +18,7 @@ extern GameObj *pInstForCollisionDetect;
 // ======================
 // 帧时间：Unity中的dt
 // ======================
-extern double frameTime;
+extern float frameTime;
 
 Status GetWinMaxMinXY();
